[#61] refactor: DB操作メソッドをプロバイダークラス化
